Minimum number of Rabbits
There are a group of N rabbits. S rabbits are chosen where S is between 0 and N inclusive. Each one of the S rabbits tells you how many other rabbits have the same color. The answers are stored in an array.

Return the minimum possible of rabbits that satisfy the given array of answers.